fix: Only append ".0" to browser version number if required
The ".0" value appears to be appended, in order for the "getSubVersion"
parser to work for certain browsers.

However, this breaks the parsing of "Samsung Internet" browsers as it
uses the "getVersionMatch" parser method.  If we append a ".0" to the
version, we don't get a version match as "isNaN(versionString)" fails.

We should only append a ".0" if the incoming browser version doesn't
contain a sub-version.

I've only added a rudimental "." string check - hopefully that's enough!